BWS-179: * make pagination for bs-space
