+++ frontend/src/components/bs-space/components/objects-table/styles.module.scss
@@ -1,0 +1,36 @@
+.buttonsBlock {
+  display: flex;
+  align-items: center;
+  gap: 14px;
+  width: 250px;
+}
+
+.hideDefaultInput {
+  position: absolute;
+  width: 0.1px;
+  height: 0.1px;
+  overflow: hidden;
+  opacity: 0;
+}
+
+.fileInput {
+  width: 100%;
+  min-height: 45px;
+  padding-top: 8px;
+  padding-bottom: 8px;
+  color: var(--color-white);
+  font-weight: 800;
+  font-size: 16px;
+  text-align: center;
+  text-decoration: none;
+  background-color: var(--color-orange);
+  border: 2px solid var(--color-orange);
+  border-radius: 50px;
+  cursor: pointer;
+  transition-duration: 0.4s;
+}
+
+.fileInput:hover {
+  color: var(--color-orange);
+  background-color: transparent;
+}
