+++ frontend/src/components/bs-space/styles.module.scss
@@ -36,40 +36,3 @@
 .addSpaceBtn {
   max-width: 200px;
 }
-
-.buttonsBlock {
-  display: flex;
-  align-items: center;
-  gap: 14px;
-  width: 250px;
-}
-
-.hideDefaultInput {
-  position: absolute;
-  width: 0.1px;
-  height: 0.1px;
-  overflow: hidden;
-  opacity: 0;
-}
-
-.fileInput {
-  width: 100%;
-  min-height: 45px;
-  padding-top: 8px;
-  padding-bottom: 8px;
-  color: var(--color-white);
-  font-weight: 800;
-  font-size: 16px;
-  text-align: center;
-  text-decoration: none;
-  background-color: var(--color-orange);
-  border: 2px solid var(--color-orange);
-  border-radius: 50px;
-  cursor: pointer;
-  transition-duration: 0.4s;
-}
-
-.fileInput:hover {
-  color: var(--color-orange);
-  background-color: transparent;
-}
